Move task-impls to hotshot
