+++ Cargo.toml
@@ -19,7 +19,6 @@ members = [
         "crates/libp2p-networking",
         "crates/testing-macros",
         "crates/task",
-        "crates/task-impls",
         "crates/testing",
         "crates/types",
         "crates/utils",
