+++ crates/hotshot/Cargo.toml
@@ -97,7 +97,6 @@ hotshot-orchestrator = { version = "0.1.1", path = "../orchestrator", default-fe
 hotshot-types = { path = "../types", version = "0.1.0", default-features = false }
 hotshot-utils = { path = "../utils" }
 hotshot-task = { path = "../task", version = "0.1.0", default-features = false }
-hotshot-task-impls = { path = "../task-impls", version = "0.1.0", default-features = false }
 libp2p-identity = { workspace = true }
 libp2p-networking = { workspace = true }
 rand = { workspace = true }
@@ -108,6 +107,8 @@ time = { workspace = true }
 derive_more = "0.99.17"
 portpicker = "0.1.1"
 lru = "0.12.1"
+bitvec = { workspace = true }
+sha2 = { workspace = true }
 
 tracing = { workspace = true }
