+++ crates/hotshot/src/tasks/HotShot_event_architecture.drawio
