+++ crates/hotshot/src/tasks/helpers.rs
