--- crates/task-impls/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-authors = ["Espresso Systems <hello@espressosys.com>"]
-description = "Async task implementations for consensus"
-edition = "2021"
-name = "hotshot-task-impls"
-version = "0.1.0"
-
-[dependencies]
-async-compatibility-layer = { workspace = true }
-async-trait = { workspace = true }
-either = { workspace = true }
-futures = { workspace = true }
-snafu = { workspace = true }
-async-lock = { workspace = true }
-tracing = { workspace = true }
-hotshot-constants = { path = "../constants", default-features = false }
-hotshot-types = { path = "../types", default-features = false }
-hotshot-task = { path = "../task", default-features = false }
-hotshot-utils = { path = "../utils" }
-time = { workspace = true }
-commit = { workspace = true }
-bincode = { workspace = true }
-bitvec = { workspace = true }
-sha2 = { workspace = true }
-
-[target.'cfg(all(async_executor_impl = "tokio"))'.dependencies]
-tokio = { workspace = true }
-[target.'cfg(all(async_executor_impl = "async-std"))'.dependencies]
-async-std = { workspace = true }
-
-[lints]
-workspace = true
