+++ crates/testing/tests/da_task.rs
@@ -1,5 +1,4 @@
-use hotshot::{types::SignatureKey, HotShotConsensusApi};
-use hotshot_task_impls::events::HotShotEvent;
+use hotshot::{tasks::events::HotShotEvent, types::SignatureKey, HotShotConsensusApi};
 use hotshot_testing::{
     block_types::TestTransaction,
     node_types::{MemoryImpl, TestTypes},
@@ -21,8 +20,7 @@ use std::{collections::HashMap, marker::PhantomData};
 )]
 #[cfg_attr(async_executor_impl = "async-std", async_std::test)]
 async fn test_da_task() {
-    use hotshot::tasks::add_da_task;
-    use hotshot_task_impls::harness::run_harness;
+    use hotshot::tasks::{add_da_task, harness::run_harness};
     use hotshot_testing::task_helpers::build_system_handle;
     use hotshot_types::message::Proposal;
